\name{logbuch-package}
\alias{logbuch-package}
\alias{logbuch}
\docType{package}
\title{
Package for handling workday log record files
}
\description{
Package for handling workday log record files
}
\details{
\tabular{ll}{
Package: \tab logbuch\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-10\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Michel Kuhlmann
}
\references{
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
